refactor: Phase 2: Complex Provider Migration
