+++ src/taskFactory.ts
@@ -4,23 +4,11 @@ import { TaskItem } from './taskItem';
 import { WorkspaceTasksService } from './services/workspaceTasksService';
 import { TaskEnvService } from './services/taskEnvService';
 import { TaskSecretWarningService } from './services/taskSecretWarningService';
-import { AntTaskProvider } from './providers/antTaskProvider';
-import { MsBuildTaskProvider } from './providers/msbuildTaskProvider';
-import { GruntTaskProvider } from './providers/gruntTaskProvider';
-import { GulpTaskProvider } from './providers/gulpTaskProvider';
-import { JustfileTaskProvider } from './providers/justfileTaskProvider';
 import { NpmTaskProvider } from './providers/npmTaskProvider';
-import { MakefileTaskProvider } from './providers/makefileTaskProvider';
-import { GithubActionsTaskProvider } from './providers/githubActionsTaskProvider';
-import { CMakeTaskProvider } from './providers/cmakeTaskProvider';
-import { CakeTaskProvider } from './providers/cakeTaskProvider';
-import { TaskfileTaskProvider } from './providers/taskfileTaskProvider';
-import { GitlabCiTaskProvider } from './providers/gitlabCiTaskProvider';
-import { CircleCiTaskProvider } from './providers/circleCiTaskProvider';
-import { BitbucketPipelinesTaskProvider } from './providers/bitbucketPipelinesTaskProvider';
 import { resolveTaskContext, CreatedTask } from './libs/taskCreationUtils';
 import { TaskProviderRegistry } from './taskProviderRegistry';
 import { LoggerService } from './services/loggerService';
+import { ensureTaskProviderRegistryPopulated } from './providers';
 
 // Re-exported for backward compatibility — CreatedTask is defined in taskCreationUtils.
 export type { CreatedTask } from './libs/taskCreationUtils';
@@ -53,53 +41,6 @@ export function injectArgs(command: string, args?: string): string {
   }
 
   return `${command} ${args}`;
-}
-
-/**
- * Collects `workflow_dispatch` input values from the user via `showInputBox` prompts.
- * Handles both the legacy `string[]` format and the current `Record<string, WorkflowInput>` format.
- * Returns an array of `--input key=value` argument pairs ready to be spread into the act args.
- */
-async function _collectWorkflowDispatchInputs(inputs: string[] | Record<string, any>): Promise<string[]> {
-  const collected: string[] = [];
-  if (Array.isArray(inputs)) {
-    for (const input of inputs) {
-      const val = await vscode.window.showInputBox({
-        prompt: `Enter input for '${input}'`,
-        placeHolder: 'Value',
-        ignoreFocusOut: true,
-      });
-      if (val) {
-        collected.push('--input', `${input}=${val}`);
-      }
-    }
-  } else {
-    for (const [key, details] of Object.entries(inputs)) {
-      const desc = (details as any).description || `Enter value for ${key}`;
-      const defaultVal = (details as any).default !== undefined ? String((details as any).default) : '';
-      const required = (details as any).required || false;
-      // type: string, boolean, choice, environment, ...
-      // For now treat all as string input
-
-      const val = await vscode.window.showInputBox({
-        prompt: desc,
-        placeHolder: `${key} (${(details as any).type || 'string'})`,
-        value: defaultVal,
-        ignoreFocusOut: true,
-        validateInput: (text) => {
-          if (required && !text) {
-            return 'This input is required';
-          }
-          return null;
-        },
-      });
-
-      if (val) {
-        collected.push('--input', `${key}=${val}`);
-      }
-    }
-  }
-  return collected;
 }
 
 /**
@@ -155,6 +96,7 @@ async function _buildTask(item: TaskItem, args?: string): Promise<CreatedTask | 
   // If the provider has implemented createTask() and it returns a non-undefined result,
   // use that. If it returns undefined (either the base-class default for not-yet-migrated
   // providers, or a legitimate "cannot build" signal), fall through to Track 2.
+  ensureTaskProviderRegistryPopulated();
   const registry = TaskProviderRegistry.getInstance();
   const registryProvider = registry.get(item.taskType);
   if (registryProvider) {
@@ -245,138 +187,6 @@ async function _buildTask(item: TaskItem, args?: string): Promise<CreatedTask | 
       );
       return { task, command: commandString, cwd, native: false };
     }
-    case 'grunt': {
-      const gruntProvider = new GruntTaskProvider();
-      const workspaceFolder = vscode.workspace.getWorkspaceFolder(resourceUri);
-      const {
-        command: gruntCmd,
-        args: gruntInitialArgs,
-        cwd: gruntCwd,
-      } = gruntProvider.getCommand(workspaceFolder?.uri);
-
-      const gruntArgs: string[] = gruntInitialArgs ? [...gruntInitialArgs] : [];
-      if (taskLabel && taskLabel.length > 0) {
-        gruntArgs.push(taskLabel);
-      }
-      if (resourceUri) {
-        const dir = path.dirname(resourceUri.fsPath);
-        const rel = path.relative(gruntCwd, dir);
-        const fileName = path.basename(resourceUri.fsPath).toLowerCase();
-
-        // If the file is not in the CWD (workspace root) or the filename is not "gruntfile.js",
-        // we need to pass the --gruntfile argument explicitly.
-        if ((rel.length > 0 && rel !== '.') || fileName !== 'gruntfile.js') {
-          gruntArgs.push('--gruntfile', resourceUri.fsPath);
-        }
-      }
-      if (args) {
-        gruntArgs.push(...args.split(' '));
-      }
-
-      const task = new vscode.Task(
-        { type: 'grunt', target: taskLabel, path: resourceUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'grunt',
-        new vscode.ShellExecution(gruntCmd, gruntArgs, { cwd: gruntCwd }),
-      );
-      return { task, command: `${gruntCmd} ${gruntArgs.join(' ')}`.trim(), cwd: gruntCwd, native: false };
-    }
-    case 'gulp': {
-      // Use npx to prefer workspace-local gulp if available
-      const gulpCmd = 'npx';
-      // We'll build the args carefully so --gulpfile appears before the task name if needed
-      let gulpArgs: string[] = ['gulp'];
-
-      // Prefer workspace folder root as cwd so local install (node_modules) is resolved correctly
-      // Use the workspace root (first workspace folder) as the default cwd so we prefer the workspace-local gulp installation
-      const defaultWorkspaceRoot =
-        vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders.length
-          ? vscode.workspace.workspaceFolders[0].uri.fsPath
-          : undefined;
-      const gulpProvider = new GulpTaskProvider();
-      const workspaceFolder = vscode.workspace.getWorkspaceFolder(resourceUri);
-      // const { command: gulpProviderCmd, cwd: gulpProviderCwd } = gulpProvider.getCommand(workspaceFolder?.uri);
-      // gulpProviderCmd is never used since we build the command manually using 'npx gulp'
-      const { command: _, cwd: gulpProviderCwd } = gulpProvider.getCommand(workspaceFolder?.uri);
-      const gulpCwd = gulpProviderCwd || defaultWorkspaceRoot || cwd;
-
-      // If the gulpfile is not located in the cwd, ensure we pass it explicitly right after 'gulp'
-      if (resourceUri && path.dirname(resourceUri.fsPath) !== gulpCwd) {
-        gulpArgs = ['gulp', '--gulpfile', resourceUri.fsPath];
-        if (taskLabel && taskLabel.length > 0) {
-          gulpArgs.push(taskLabel);
-        }
-      } else {
-        if (taskLabel && taskLabel.length > 0) {
-          gulpArgs.push(taskLabel);
-        }
-      }
-
-      if (args) {
-        gulpArgs.push(...args.split(' '));
-      }
-
-      const task = new vscode.Task(
-        { type: 'gulp', target: taskLabel, path: resourceUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'gulp',
-        new vscode.ShellExecution(gulpCmd, gulpArgs, { cwd: gulpCwd }),
-      );
-
-      return { task, command: `${gulpCmd} ${gulpArgs.join(' ')}`.trim(), cwd: gulpCwd, native: false };
-    }
-    case 'ant': {
-      const antProvider = new AntTaskProvider();
-      const useAnsicon = antProvider.shouldUseAnsicon();
-      const workspaceFolder = vscode.workspace.getWorkspaceFolder(resourceUri);
-      const workspaceUri = workspaceFolder?.uri;
-
-      const { command: antCommand, args: antInitialArgs, cwd: antCwd } = antProvider.getCommand(workspaceUri);
-
-      let command: string;
-      let commandArgs: string[];
-
-      if (useAnsicon) {
-        command = antProvider.getAnsicon(workspaceUri).command;
-        commandArgs = [antCommand];
-        if (antInitialArgs) {
-          commandArgs.push(...antInitialArgs);
-        }
-        commandArgs = commandArgs.concat(antProvider.getCommandArgs(taskLabel, true, item.taskFileUri?.fsPath));
-      } else {
-        command = antCommand;
-        commandArgs = antInitialArgs ? [...antInitialArgs] : [];
-        commandArgs.push(...antProvider.getCommandArgs(taskLabel, false, item.taskFileUri?.fsPath));
-      }
-
-      if (args) {
-        commandArgs.push(...args.split(' '));
-      }
-
-      if (useAnsicon) {
-        const ansiconArgs = commandArgs.map((arg) => `"${arg}"`).join(' ');
-        const fullCommand = `"${command}" ${ansiconArgs}`;
-        const task = new vscode.Task(
-          { type: 'ant', target: taskLabel, path: resourceUri.fsPath },
-          vscode.TaskScope.Workspace,
-          taskLabel,
-          'ant',
-          new vscode.ShellExecution(fullCommand, { cwd: antCwd }),
-        );
-        return { task, command: fullCommand, cwd: antCwd, native: false };
-      } else {
-        const task = new vscode.Task(
-          { type: 'ant', target: taskLabel, path: resourceUri.fsPath },
-          vscode.TaskScope.Workspace,
-          taskLabel,
-          'ant',
-          new vscode.ShellExecution(command, commandArgs, { cwd: antCwd }),
-        );
-        return { task, command: `${command} ${commandArgs.join(' ')}`, cwd: antCwd, native: false };
-      }
-    }
     case 'workspace-task': {
       // This case occurs when a workspace task exists without file association
       const configType = item.taskSource || 'shell';
@@ -393,116 +203,6 @@ async function _buildTask(item: TaskItem, args?: string): Promise<CreatedTask | 
         return { task, command: fullCommand, cwd, native: false };
       }
       return undefined;
-    }
-    case 'github-actions': {
-      const ghProvider = new GithubActionsTaskProvider();
-      const workspaceFolder = vscode.workspace.getWorkspaceFolder(resourceUri);
-      let { command: actPath, args: actInitialArgs, cwd: actCwd } = ghProvider.getCommand(workspaceFolder?.uri);
-
-      // Attempt to determine the project root if the workflow is in a nested folder
-      // This is crucial for act to find .secrets, .env in the project root instead of workspace root
-      const githubDirMatch = resourceUri.fsPath.match(/[\\/]\.github[\\/]/);
-      if (githubDirMatch) {
-        const projectRoot = resourceUri.fsPath.substring(0, githubDirMatch.index);
-        if (projectRoot) {
-          actCwd = projectRoot;
-        }
-      }
-
-      const meta = item.metadata;
-      const actArgs: string[] = actInitialArgs ? [...actInitialArgs] : [];
-      const config = vscode.workspace.getConfiguration('workspaceTasks');
-      // workspaceFolder already defined above
-      // const workspaceFolder = vscode.workspace.getWorkspaceFolder(resourceUri);
-
-      // this should be the first argument.
-      if (meta?.type === 'workflow') {
-        if (meta.event === 'workflow_dispatch' && meta.inputs) {
-          // Handle both old string[] interface (fallback) and new Record interface
-          actArgs.push(...await _collectWorkflowDispatchInputs(meta.inputs as string[] | Record<string, any>));
-          actArgs.push('workflow_dispatch');
-        } else {
-          actArgs.push(meta.event || 'push');
-        }
-
-        // Handle Env File
-        let envFile = config.get<string>('act.envFile');
-        if (envFile) {
-          actArgs.push('--env-file', envFile as string);
-        }
-
-        // Handle Variables File
-        let varsFile = config.get<string>('act.variablesFile');
-        if (varsFile) {
-          actArgs.push('--var-file', varsFile as string);
-        }
-
-        // Handle Secrets File
-        let secretsFile = config.get<string>('act.secretsFile');
-        if (secretsFile) {
-          actArgs.push('--secret-file', secretsFile as string);
-        }
-
-        // Handle Variables
-        const vars = config.get<Record<string, string>>('act.variables');
-        if (vars) {
-          for (const [key, value] of Object.entries(vars)) {
-            actArgs.push('--var', `${key}=${value}`);
-          }
-        }
-
-        const relPath = path.relative(actCwd, resourceUri.fsPath);
-        actArgs.push('-W', relPath);
-      } else if (meta?.type === 'job') {
-        actArgs.push('-j', meta.jobId);
-        const relPath = path.relative(actCwd, resourceUri.fsPath);
-        actArgs.push('-W', relPath);
-      } else {
-        // Fallback or generic file execution
-        let useEvent = 'push';
-
-        // Check for file-level metadata with supported events
-        if (meta?.type === 'file' && meta?.events) {
-          const events = meta.events as string[];
-          if (events.length > 0) {
-            const selected = await vscode.window.showQuickPick(events, {
-              placeHolder: 'Select event to trigger',
-            });
-            if (selected) {
-              useEvent = selected;
-            } else {
-              // User cancelled selection
-              return undefined;
-            }
-          }
-        }
-
-        // If selected event is workflow_dispatch, handle inputs
-        if (useEvent === 'workflow_dispatch' && meta?.inputs) {
-          actArgs.push(...await _collectWorkflowDispatchInputs(meta.inputs as string[] | Record<string, any>));
-        }
-
-        actArgs.push(useEvent);
-        const relPath = path.relative(actCwd, resourceUri.fsPath);
-        actArgs.push('-W', relPath);
-      }
-
-      if (args) {
-        actArgs.push(...args.split(' '));
-      }
-
-      const task = new vscode.Task(
-        { type: 'github-actions', task: taskLabel, path: resourceUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'github-actions',
-        new vscode.ShellExecution(actPath, actArgs, { cwd: actCwd }),
-      );
-      // Reconstruct command string for display/logging purposes mostly
-      const safeCmd = /\s/.test(actPath) ? `"${actPath}"` : actPath;
-      const full = `${safeCmd} ${actArgs.map((a) => (/\s/.test(a) ? `"${a}"` : a)).join(' ')}`;
-
-      return { task, command: full, cwd: actCwd, native: false };
     }
     case 'vscode': {
       // Use existing Visual Studio Code task defined in .vscode/tasks.json
@@ -532,33 +232,6 @@ async function _buildTask(item: TaskItem, args?: string): Promise<CreatedTask | 
         return { task: found, cwd: undefined, native: true };
       }
       return undefined;
-    }
-    case 'makefile': {
-      const makeProvider = new MakefileTaskProvider();
-      const workspaceFolder = vscode.workspace.getWorkspaceFolder(resourceUri);
-      // cwd needs to be the workspace folder where Makefile is located
-      // use resourceUri to find the makefile location if possible
-      const makeFileCwd = path.dirname(resourceUri.fsPath);
-
-      const { command: makeCmd, args: makeInitialArgs } = makeProvider.getCommand(workspaceFolder?.uri);
-
-      const makeArgs = makeInitialArgs ? [...makeInitialArgs] : [];
-      makeArgs.push(taskLabel);
-      if (args) {
-        makeArgs.push(...args.split(' '));
-      }
-
-      const full = `${makeCmd} ${makeArgs.join(' ')}`;
-      const shellExec = new vscode.ShellExecution(makeCmd, makeArgs, { cwd: makeFileCwd });
-
-      const task = new vscode.Task(
-        { type: 'process', script: taskLabel, path: resourceUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'makefile',
-        shellExec,
-      );
-      return { task, command: full, cwd: makeFileCwd, native: false };
     }
     case 'dockerfile': {
       const command = await WorkspaceTasksService.getInstance().resolveTaskCommand(
@@ -614,297 +287,6 @@ async function _buildTask(item: TaskItem, args?: string): Promise<CreatedTask | 
         shellExec,
       );
       return { task, command: commandString, cwd, native: false };
-    }
-    case 'msbuild': {
-      // item.taskFileUri is the actual project file URI. item.resourceUri is always a
-      // synthetic workspace-tasks:// URI set by updateContextValue(), so it cannot be
-      // used as a guard. When no real project file is associated, return undefined
-      // instead of falling through with an invalid MSBuild invocation.
-      if (!effectiveResourceUri) {
-        return undefined;
-      }
-      const msbuildProvider = new MsBuildTaskProvider();
-      const workspaceFolder = vscode.workspace.getWorkspaceFolder(resourceUri);
-      const workspaceUri = workspaceFolder?.uri;
-      const {
-        command: msbuildCmd,
-        args: msbuildInitialArgs,
-        cwd: msbuildCwd,
-      } = msbuildProvider.getCommand(workspaceUri);
-      const commandArgs = msbuildInitialArgs ? [...msbuildInitialArgs] : [];
-      commandArgs.push(...msbuildProvider.getCommandArgs(taskLabel, resourceUri.fsPath));
-      if (args) {
-        commandArgs.push(...args.split(' '));
-      }
-      const task = new vscode.Task(
-        { type: 'msbuild', target: taskLabel, path: resourceUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'msbuild',
-        new vscode.ShellExecution(msbuildCmd, commandArgs, { cwd: msbuildCwd }),
-      );
-      return { task, command: `${msbuildCmd} ${commandArgs.join(' ')}`, cwd: msbuildCwd, native: false };
-    }
-    case 'justfile': {
-      const justProvider = new JustfileTaskProvider();
-      const { command: justCommand, args: justInitialArgs, cwd: justCwd } = justProvider.getCommand(resourceUri);
-
-      const justArgs = justInitialArgs ? [...justInitialArgs] : [];
-      justArgs.push('--justfile', resourceUri.fsPath);
-      justArgs.push(taskLabel);
-      if (args) {
-        justArgs.push(...args.split(' '));
-      }
-
-      const fullCmd = `${justCommand} ${justArgs.join(' ')}`.trim();
-
-      const task = new vscode.Task(
-        { type: 'justfile', task: taskLabel, path: resourceUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'just',
-        new vscode.ShellExecution(justCommand, justArgs, { cwd: justCwd }),
-      );
-      return { task, command: fullCmd, cwd: justCwd, native: false };
-    }
-    case 'cmake': {
-      const cmakeProvider = new CMakeTaskProvider();
-      const workspaceFolder = vscode.workspace.getWorkspaceFolder(resourceUri);
-      const { command: cmakeCmd, args: cmakeInitialArgs } = cmakeProvider.getCommand(workspaceFolder?.uri);
-
-      const sourceDir = path.dirname(resourceUri.fsPath);
-      const buildDir = cmakeProvider.getBuildDirectory();
-      const buildPath = path.isAbsolute(buildDir) ? buildDir : path.join(sourceDir, buildDir);
-      const buildType = cmakeProvider.getBuildType();
-
-      const cmakeArgs = cmakeInitialArgs ? [...cmakeInitialArgs] : [];
-      cmakeArgs.push('--build', buildPath, '--target', taskLabel, '--config', buildType);
-
-      if (args) {
-        cmakeArgs.push(...args.split(' '));
-      }
-
-      const full = `${cmakeCmd} ${cmakeArgs.join(' ')}`;
-      const shellExec = new vscode.ShellExecution(cmakeCmd, cmakeArgs, { cwd: sourceDir });
-
-      const task = new vscode.Task(
-        { type: 'cmake', target: taskLabel, path: resourceUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'cmake',
-        shellExec,
-      );
-      return { task, command: full, cwd: sourceDir, native: false };
-    }
-    case 'cake': {
-      const cakeProvider = new CakeTaskProvider();
-      const workspaceFolder = vscode.workspace.getWorkspaceFolder(resourceUri);
-      const { command: cakeCmd, args: cakeInitialArgs, cwd: cakeCwd } = cakeProvider.getCommand(workspaceFolder?.uri);
-
-      const cakeArgs = cakeInitialArgs ? [...cakeInitialArgs] : [];
-      cakeArgs.push(resourceUri.fsPath, `--target=${taskLabel}`);
-      if (args) {
-        cakeArgs.push(...args.split(' '));
-      }
-
-      const full = `${cakeCmd} ${cakeArgs.join(' ')}`;
-      const shellExec = new vscode.ShellExecution(cakeCmd, cakeArgs, { cwd: cakeCwd });
-
-      const task = new vscode.Task(
-        { type: 'cake', target: taskLabel, path: resourceUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'cake',
-        shellExec,
-      );
-      return { task, command: full, cwd: cakeCwd, native: false };
-    }
-    case 'taskfile': {
-      const taskfileProvider = new TaskfileTaskProvider();
-      const workspaceFolder = vscode.workspace.getWorkspaceFolder(resourceUri);
-      const { command: taskCmd, args: taskInitialArgs, cwd: taskCwd } = taskfileProvider.getCommand(workspaceFolder?.uri);
-
-      const taskArgs = taskInitialArgs ? [...taskInitialArgs] : [];
-      const globalTaskfilePath = typeof item.metadata?.globalTaskfilePath === 'string'
-        ? item.metadata.globalTaskfilePath
-        : undefined;
-
-      if (globalTaskfilePath) {
-        taskArgs.push('--taskfile', globalTaskfilePath);
-      } else if (item.taskFileUri) {
-        taskArgs.push('--taskfile', item.taskFileUri.fsPath);
-      }
-      taskArgs.push(taskLabel);
-      if (args) {
-        taskArgs.push(...args.split(' '));
-      }
-
-      const full = `${taskCmd} ${taskArgs.join(' ')}`;
-      const taskFileCwd = item.taskFileUri ? path.dirname(item.taskFileUri.fsPath) : taskCwd;
-      const shellExec = new vscode.ShellExecution(taskCmd, taskArgs, { cwd: taskFileCwd });
-
-      const task = new vscode.Task(
-        { type: 'taskfile', task: taskLabel, path: resourceUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'taskfile',
-        shellExec,
-      );
-      return { task, command: full, cwd: taskFileCwd, native: false };
-    }
-    case 'gitlab-ci': {
-      if (!item.taskFileUri) {
-        return undefined;
-      }
-      const ciProvider = new GitlabCiTaskProvider();
-      const { command: ciCmd, args: providerArgs } = ciProvider.getCommand(item.taskFileUri);
-      const ciArgs = [...(providerArgs ?? [])];
-
-      const effectiveFileUri = item.taskFileUri;
-      ciArgs.push('--file', path.basename(effectiveFileUri.fsPath));
-
-      const ciConfig = vscode.workspace.getConfiguration('workspaceTasks');
-      const variablesFile = ciConfig.get<string>('gitlabCiLocal.variablesFile', '');
-      if (variablesFile) {
-        ciArgs.push('--variables-file', variablesFile);
-      }
-      const variables = ciConfig.get<string[]>('gitlabCiLocal.variable', []);
-      for (const v of variables) {
-        ciArgs.push('--variable', v);
-      }
-      const unsetVars = ciConfig.get<string[]>('gitlabCiLocal.unsetVariable', []);
-      for (const u of unsetVars) {
-        ciArgs.push('--unset-variable', u);
-      }
-      const remoteVars = ciConfig.get<string[]>('gitlabCiLocal.remoteVariables', []);
-      for (const r of remoteVars) {
-        ciArgs.push('--remote-variables', r);
-      }
-      const home = ciConfig.get<string>('gitlabCiLocal.home', '');
-      if (home) {
-        ciArgs.push('--home', home);
-      }
-
-      ciArgs.push(taskLabel);
-      if (args) {
-        ciArgs.push(...args.split(' '));
-      }
-
-      const ciCwd = path.dirname(effectiveFileUri.fsPath);
-      const shellExec = new vscode.ShellExecution(ciCmd, ciArgs, { cwd: ciCwd });
-      const full = `${ciCmd} ${ciArgs.join(' ')}`;
-
-      const task = new vscode.Task(
-        { type: 'gitlab-ci', job: taskLabel, path: effectiveFileUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'gitlab-ci',
-        shellExec,
-      );
-      return { task, command: full, cwd: ciCwd, native: false };
-    }
-    case 'circleci': {
-      if (item.metadata?.type && item.metadata.type !== 'job') {
-        return undefined;
-      }
-      if (!item.taskFileUri) {
-        return undefined;
-      }
-
-      const circleProvider = new CircleCiTaskProvider();
-      const { command: circleCmd, args: providerArgs } = circleProvider.getCommand(item.taskFileUri);
-      const circleArgs = [...(providerArgs ?? [])];
-
-      const circleWorkspaceFolder = vscode.workspace.getWorkspaceFolder(item.taskFileUri);
-      const circleCwd = circleWorkspaceFolder?.uri.fsPath ?? path.dirname(item.taskFileUri.fsPath);
-      const configPath = circleWorkspaceFolder
-        ? path.relative(circleWorkspaceFolder.uri.fsPath, item.taskFileUri.fsPath) || path.basename(item.taskFileUri.fsPath)
-        : item.taskFileUri.fsPath;
-
-      circleArgs.push(
-        'local',
-        'execute',
-        '-c',
-        configPath,
-        taskLabel,
-      );
-
-      if (args) {
-        circleArgs.push(...args.split(' '));
-      }
-
-      const shellExec = new vscode.ShellExecution(circleCmd, circleArgs, { cwd: circleCwd });
-      const full = `${circleCmd} ${circleArgs.join(' ')}`;
-
-      const task = new vscode.Task(
-        { type: 'circleci', job: taskLabel, path: item.taskFileUri.fsPath },
-        vscode.TaskScope.Workspace,
-        taskLabel,
-        'circleci',
-        shellExec,
-      );
-      return { task, command: full, cwd: circleCwd, native: false };
-    }
-    case 'bitbucket': {
-      if (!item.taskFileUri) {
-        return undefined;
-      }
-
-      const bbProvider = new BitbucketPipelinesTaskProvider();
-      const { command: bbCmd, args: providerArgs } = bbProvider.getCommand(item.taskFileUri);
-      const bbArgs = [...(providerArgs ?? [])];
-      bbArgs.push('run');
-
-      const bbConfig = vscode.workspace.getConfiguration('workspaceTasks');
-      const envFiles = bbConfig.get<string[]>('bitbucketPipelineRunner.environmentFiles', []);
-      for (const ef of envFiles) {
-        bbArgs.push('--env-file', ef);
-      }
-
-      const meta = item.metadata;
-      if (meta?.type === 'step') {
-        if (typeof meta.stepName !== 'string' || typeof meta.pipelinePath !== 'string') {
-          return undefined;
-        }
-        bbArgs.push('--step', meta.stepName);
-        bbArgs.push(meta.pipelinePath);
-      } else if (meta?.type === 'stage') {
-        if (typeof meta.stageName !== 'string' || typeof meta.pipelinePath !== 'string') {
-          return undefined;
-        }
-        bbArgs.push('--stage', meta.stageName);
-        bbArgs.push(meta.pipelinePath);
-      } else if (meta?.type === 'pipeline') {
-        if (typeof meta.pipelinePath !== 'string') {
-          return undefined;
-        }
-        bbArgs.push(meta.pipelinePath);
-      } else {
-        return undefined;
-      }
-
-      if (args) {
-        bbArgs.push(...args.split(' '));
-      }
-
-      const bbWorkspaceFolder = vscode.workspace.getWorkspaceFolder(item.taskFileUri);
-      const bbCwd = bbWorkspaceFolder?.uri.fsPath ?? path.dirname(item.taskFileUri.fsPath);
-
-      const shellExec = new vscode.ShellExecution(bbCmd, bbArgs, { cwd: bbCwd });
-      const full = `${bbCmd} ${bbArgs.map((a) => (/\s/.test(a) ? `"${a}"` : a)).join(' ')}`;
-
-      const taskDef: { type: string; pipeline: string; stage?: string; step?: string } = {
-        type: 'bitbucket',
-        pipeline: meta.pipelinePath as string,
-      };
-      if (meta.type === 'stage') {
-        taskDef.stage = meta.stageName as string;
-      } else if (meta.type === 'step') {
-        taskDef.step = meta.stepName as string;
-      }
-
-      const bbTask = new vscode.Task(taskDef, vscode.TaskScope.Workspace, taskLabel, 'bitbucket', shellExec);
-      return { task: bbTask, command: full, cwd: bbCwd, native: false };
     }
     default: {
       // Generic: run as shell command if workspace has a declared task
